feat property search state
