+++ src/App.js
@@ -10,37 +10,12 @@ import Footer from "./Components/Footer";
 import Testimonials from "./Components/Testimonials";
 import { useState, useEffect } from "react";
 function App() {
-  const [propertiesList, setPropertiesList] = useState(null);
-  
-  
-
-  const getData = () => {
-    fetch("data.json", {
-      headers: {
-        "Content-Type": "application/json",
-        Accept: "application/json",
-      },
-    })
-      .then(function (response) {
-        return response.json();
-      })
-      .then(function (myJson) {
-        setPropertiesList(myJson);
-      });
-  };
-
-  useEffect(() => {
-    getData();
-    
-  }, []);
-
-  
 
   return (
     <div className="">
       <Header />
       <Benefits />
-      <Properties propertiesList={propertiesList}/>
+      <Properties />
       <Flexibility />
       <AddProperty />
       <Testimonials />
